requirements.txt
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
langchain_cohere
InstructorEmbedding